mlops-multi-account-cdk/mlops-sm-project-template/requirements.txt
aws-cdk-lib
awscliv2
boto3
constructs
dataclasses-json
jupyter
pre-commit
yamldataclassconfig
